+++ app/src/main/java/com/jin/jjinweather/feature/location/LocationRepository.kt
@@ -1,4 +1,4 @@
-package com.jin.jjinweather.layer.domain.repository
+package com.jin.jjinweather.feature.location
 
 import com.jin.jjinweather.layer.domain.model.location.GeoPoint
